Split interface and implementation in maps package.
